payment-admin/pyproject.toml
[tool.poetry]
name = "payment-admin"
version = "0.1.0"
description = "Admin panel for payments and subscriptions"
authors = ["Vasiliy Ivanov <shitcoder@pm.me>"]

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.4"
psycopg2-binary = "^2.9.9"
django-split-settings = "^1.3.1"
python-dotenv = "^1.0.1"
circuitbreaker = "^2.0.0"
backoff = "^2.2.1"
django-postgres-extra = "^2.0.8"
uwsgi = "^2.0.25.1"
requests = "^2.31.0"
django-currentuser = "^0.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
